David did this
I've done the buttons logic. It should be working just as well. I've also made a couple of changes to the map and i just styled the navigation bar a little bit. It might need a little bit of styling on the <divs> with the main content. We should not make a whole app, just a demo. We can leave those tiny pieces for the main event.